Ticket #6: Remove Contest + Update Carousel Item
Remove the whole contest area of the page, everything from below the black box containing the products should be removed. We can recover this via SVN later, just make sure to add this as a comment in the commit so we can retrieve it. 
Also, replace the first carousel images being used on the english and french pages with the ones attached to this ticket outlining that the contest is "coming soon"

ATTACHMENTS:
1.	newheader_carousel_item_EN.png (198.27kb, image/png)
2.	newheader_carousel_item_FR.png (203.16kb, image/png)